Рефакторинг проекта
