+++ utils.php
@@ -1,0 +1,53 @@
+<?php
+    function include_template($name, $data) {
+        $name = 'templates/' . $name;
+        $result = '';
+
+        if (!is_readable($name)) {
+            return $result;
+        }
+
+        ob_start();
+        extract($data);
+        require $name;
+
+        $result = ob_get_clean();
+
+        return $result;
+    }
+
+    function db_get_prepare_stmt($link, $sql, $data = []) {
+        $stmt = mysqli_prepare($link, $sql);
+
+        if ($data) {
+            $types = '';
+            $stmt_data = [];
+
+            foreach ($data as $value) {
+                $type = null;
+
+                if (is_int($value)) {
+                    $type = 'i';
+                }
+                else if (is_string($value)) {
+                    $type = 's';
+                }
+                else if (is_double($value)) {
+                    $type = 'd';
+                }
+
+                if ($type) {
+                    $types .= $type;
+                    $stmt_data[] = $value;
+                }
+            }
+
+            $values = array_merge([$stmt, $types], $stmt_data);
+
+            $func = 'mysqli_stmt_bind_param';
+            $func(...$values);
+        }
+
+        return $stmt;
+    }
+
